removed then, after await  corrected name of function and modified setshow as per review

+++ jiraworklog/src/Container/Login/Login.js
@@ -1,18 +1,18 @@
 import React, { Component } from 'react';
 import { withRouter } from 'react-router-dom';
 import { Button, Grid, Header, Segment, Message } from 'semantic-ui-react'
-import { Form} from 'semantic-ui-react-form-validator';
+import { Form } from 'semantic-ui-react-form-validator';
 import Api from '../../utility';
-import {emailObject,passwordObject,urlObject} from '../../Constants';
+import { emailObject, passwordObject, urlObject } from '../../Constants';
 import InputHoc from '../../Hoc/FormInput/input';
 
 class Login extends Component {
-state = {
-  email: "",
-  apitoken: "",
-  apiurl: "",
-  errorMsg: false
-}
+  state = {
+    email: "",
+    apitoken: "",
+    apiurl: "",
+    errorMsg: false
+  }
 
   fetchData = async () => {
     const { email, apitoken, apiurl } = this.state;
@@ -21,29 +21,26 @@ state = {
     localStorage.setItem('email', email)
 
     try {
-       await Api.apicall(`${apiurl}/rest/api/2/project`)
-      .then(response=>
-        { if (response.length > 0){
-          localStorage.clear();
-          localStorage.setItem('apiToken', apitoken)
-          localStorage.setItem('url', apiurl)
-          localStorage.setItem('email', email)
-          this.props.history.push('/timesheet')
-
-        }}
-      )
-      
+      const response = await Api.apicall(`${apiurl}/rest/api/2/project`)
+      if (response.length > 0) {
+        localStorage.clear();
+        localStorage.setItem('apiToken', apitoken)
+        localStorage.setItem('url', apiurl)
+        localStorage.setItem('email', email)
+        this.props.history.push('/timesheet')
+      }
     } catch (error) {
       this.setState({ errorMsg: true })
     }
   }
 
-  onChangehandler=(inputValue,event)=>{
-    this.setState({ [inputValue]: event.target.value }) }
+  onChangehandler = (inputValue, event) => {
+    this.setState({ [inputValue]: event.target.value })
+  }
 
   render() {
-    const {errorMsg,email,apitoken,apiurl } = this.state;
-    const {inputs}=this.props;
+    const { errorMsg, email, apitoken, apiurl } = this.state;
+    const { inputs } = this.props;
     const LoginForm =
       <Grid textAlign='center' style={{ height: '100vh' }} verticalAlign='middle'>
         <Grid.Column style={{ maxWidth: 450 }}>
@@ -53,9 +50,9 @@ state = {
           <Form size='large'
             onSubmit={this.fetchData}>
             <Segment stacked>
-             {inputs({...emailObject,onChange:(event)=> this.onChangehandler("email", event),value:email})}
-              {inputs({...urlObject,onChange:(event)=>this.onChangehandler("apiurl",event),value:apiurl})}
-              {inputs({...passwordObject,onChange:(event)=>this.onChangehandler("apitoken",event),value:apitoken})}
+              {inputs({ ...emailObject, onChange: (event) => this.onChangehandler("email", event), value: email })}
+              {inputs({ ...urlObject, onChange: (event) => this.onChangehandler("apiurl", event), value: apiurl })}
+              {inputs({ ...passwordObject, onChange: (event) => this.onChangehandler("apitoken", event), value: apitoken })}
               {errorMsg ? <Message>Credientials not correct</Message> : null}
 
               <Button color='blue' fluid size='large'>
@@ -73,4 +70,4 @@ state = {
   }
 }
 
-export default  InputHoc(withRouter(Login));
+export default InputHoc(withRouter(Login));
